[fix] 그룹러닝 업데이트시 기존이미지와 새 이미지 비교 / 업데이트 로직 수정

+++ middlewares/multers/multer.js
@@ -80,15 +80,16 @@ const uploadProfile = multer({
 })
 
 const deleteImg = (url) => {
-    const filename = url.split('/')[4]
-
-    s3.deleteObject(
-        {
-            Bucket: process.env.AWS_S3_BUCKET,
-            Key: filename,
-        },
-        function (err, data) {}
-    )
+    if (url) {
+        const filename = url.split('/')[4]
+        s3.deleteObject(
+            {
+                Bucket: process.env.AWS_S3_BUCKET,
+                Key: filename,
+            },
+            function (err, data) {}
+        )
+    }
 }
 
 const deleteProfile = (url) => {
